# docs/openapi.yaml
openapi: 3.1.0
info:
  title: AKSIS Core API
  version: 1.13.0
  description: >-
    Integrated Phase 03-15 REST contract covering authentication and tenant
    context, schools/classes/students, academic periods and student history,
    card lifecycle, and secure device provisioning/runtime operations.
  x-aksis-phases:
    - phase: 3
      resources: [Auth, Schools, Classes, Students]
    - phase: 4
      resources: [Academic Years, Student History, Cards]
    - phase: 5
      resources: [Devices]
    - phase: 6
      resources: [Gate Attendance, Offline Sync]
    - phase: 7
      resources: [Admin Context, Attendance Read Model, Device Inventory]
    - phase: 8
      resources: [Parent Linking, Children, Parent Timeline]
    - phase: 9
      resources: [Waste Transactions, Ranking, Summary]
    - phase: 10
      resources: [Extracurriculars, Sessions, Members, Attendance]
    - phase: 11
      resources: [Library Visits, Offline Sync, Class Summary]
    - phase: 12
      resources: [LED Content, Gateway State, Heartbeat, Acknowledgement]
    - phase: 13
      resources: [Command Center, Today Aggregate]
    - phase: 14
      resources: [Card Writer Jobs, Station Lease, Read-back Verification]
    - phase: 15
      resources: [Reporting, Audit, Monitoring]
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
...
security:
  - bearerAuth: []
    schoolId: []
paths:
...
  /devices:
    get:
      tags: [Devices]
      operationId: listDevices
      summary: List the selected school's device inventory
      description: Requires `device.read`; credential material is never returned.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Search'
      responses:
        '200': { description: Paginated device inventory, content: { application/json: { schema: { $ref: '#/components/schemas/DeviceListResponse' } } } }
        '401': { $ref: '#/components/responses/AuthRequired' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /platform/stats:
    get:
      tags: [Platform]
      operationId: platformStats
      summary: "Aggregate platform metrics; requires trusted platform authority"
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: Success envelope }, '403': { $ref: '#/components/responses/Forbidden' }, '422': { $ref: '#/components/responses/ValidationError' } }
  /platform/schools:
    get:
      tags: [Platform]
      operationId: platformSchools
      summary: "List all non-deleted schools"
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: Success envelope }, '403': { $ref: '#/components/responses/Forbidden' }, '422': { $ref: '#/components/responses/ValidationError' } }
    post:
      tags: [Platform]
      operationId: platformProvisionSchool
      summary: "Atomically provision school and administrator; same operation key resumes an interrupted request"
      security: [{ bearerAuth: [] }]
      requestBody: {"required": true, "content": {"application/json": {"schema": {"type": "object", "additionalProperties": false, "required": ["code", "name", "timezone", "idempotency_key"], "properties": {"code": {"type": "string", "pattern": "^[A-Z0-9][A-Z0-9_-]{1,49}$"}, "name": {"type": "string"}, "timezone": {"type": "string", "enum": ["Asia/Jakarta", "Asia/Makassar", "Asia/Jayapura"]}, "idempotency_key": {"type": "string", "format": "uuid"}}}}}}
      responses: { '201': { description: Success envelope }, '403': { $ref: '#/components/responses/Forbidden' }, '422': { $ref: '#/components/responses/ValidationError' } }
  /platform/schools/{id}:
    parameters: [{ $ref: '#/components/parameters/ResourceId' }]
    patch:
      tags: [Platform]
      operationId: platformUpdateSchool
      summary: "Update school profile and status"
      security: [{ bearerAuth: [] }]
      requestBody: {"required": true, "content": {"application/json": {"schema": {"type": "object", "additionalProperties": false, "required": ["name", "timezone", "status"], "properties": {"name": {"type": "string"}, "timezone": {"type": "string", "enum": ["Asia/Jakarta", "Asia/Makassar", "Asia/Jayapura"]}, "status": {"type": "string", "enum": ["ACTIVE", "SUSPENDED", "INACTIVE"]}}}}}}
      responses: { '200': { description: Success envelope }, '403': { $ref: '#/components/responses/Forbidden' }, '422': { $ref: '#/components/responses/ValidationError' } }
  /platform/schools/{id}/admin:
    parameters: [{ $ref: '#/components/parameters/ResourceId' }]
    post:
      tags: [Platform]
      operationId: platformProvisionAdmin
      summary: "Provision an administrator in the target school"
      security: [{ bearerAuth: [] }]
      requestBody: {"required": true, "content": {"application/json": {"schema": {"type": "object", "additionalProperties": false, "required": ["email", "password", "full_name", "idempotency_key"], "properties": {"email": {"type": "string", "format": "email"}, "password": {"type": "string", "minLength": 12, "maxLength": 128}, "full_name": {"type": "string"}, "idempotency_key": {"type": "string", "format": "uuid"}}}}}}
      responses: { '201': { description: Success envelope }, '403': { $ref: '#/components/responses/Forbidden' }, '422': { $ref: '#/components/responses/ValidationError' } }
  /platform/iam:
    get:
      tags: [Platform]
      operationId: platformMemberships
      summary: "Paginated memberships with actual roles and permissions"
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: Success envelope }, '403': { $ref: '#/components/responses/Forbidden' }, '422': { $ref: '#/components/responses/ValidationError' } }
  /platform/iam/{id}:
    parameters: [{ $ref: '#/components/parameters/ResourceId' }]
    patch:
      tags: [Platform]
      operationId: platformUpdateMembership
      summary: "Update membership; cannot remove last active school administrator"
      security: [{ bearerAuth: [] }]
      requestBody: {"required": true, "content": {"application/json": {"schema": {"type": "object", "additionalProperties": false, "required": ["status", "roles"], "properties": {"status": {"type": "string", "enum": ["ACTIVE", "SUSPENDED", "REVOKED"]}, "roles": {"type": "array", "minItems": 1, "items": {"type": "string", "enum": ["SCHOOL_ADMIN", "TEACHER", "EXTRA_TEACHER", "LIBRARY_STAFF", "WASTE_STAFF", "PARENT"]}}}}}}}
      responses: { '200': { description: Success envelope }, '403': { $ref: '#/components/responses/Forbidden' }, '422': { $ref: '#/components/responses/ValidationError' } }
  /platform/devices:
    get:
      tags: [Platform]
      operationId: platformDevices
      summary: "Paginated device status; online uses a two-minute heartbeat window"
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: Success envelope }, '403': { $ref: '#/components/responses/Forbidden' }, '422': { $ref: '#/components/responses/ValidationError' } }
  /platform/devices/{id}:
    parameters: [{ $ref: '#/components/parameters/ResourceId' }]
    patch:
      tags: [Platform]
      operationId: platformDeviceStatus
      summary: "Set administrative device status"
      security: [{ bearerAuth: [] }]
      requestBody: {"required": true, "content": {"application/json": {"schema": {"type": "object", "additionalProperties": false, "required": ["status"], "properties": {"status": {"type": "string", "enum": ["ACTIVE", "DISABLED", "MAINTENANCE", "RETIRED"]}}}}}}
      responses: { '200': { description: Success envelope }, '403': { $ref: '#/components/responses/Forbidden' }, '422': { $ref: '#/components/responses/ValidationError' } }
  /platform/devices/{id}/credential:
    parameters: [{ $ref: '#/components/parameters/ResourceId' }]
    post:
      tags: [Platform]
      operationId: platformRotateDeviceCredential
      summary: "Revoke old credentials and return new token once"
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: Success envelope }, '403': { $ref: '#/components/responses/Forbidden' }, '422': { $ref: '#/components/responses/ValidationError' } }
  /platform/audit:
    get:
      tags: [Platform]
      operationId: platformAudit
      summary: "Paginated cross-school domain audit trail"
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: Success envelope }, '403': { $ref: '#/components/responses/Forbidden' }, '422': { $ref: '#/components/responses/ValidationError' } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    schoolId:
      type: apiKey
      in: header
      name: X-School-Id
      description: UUID of the active school membership.
    deviceAuth:
      type: apiKey
      in: header
      name: Authorization
      description: Per-device credential formatted as `Device <token>`.
    deviceId:
      type: apiKey
      in: header
      name: X-Device-Id
      description: Registered device UUID; school context is derived server-side.
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Search:
      name: search
      in: query
      schema: { type: string, minLength: 1, maxLength: 100, pattern: '^[\p{L}\p{N}\s.-]+$' }
  responses:
...
    AuthRequired:
      description: Missing, expired, or invalid bearer token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    Forbidden:
      description: Tenant membership or permission is missing
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    ValidationError:
      description: Request payload, parameters, or query are invalid
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
  schemas:
...
    SuccessEnvelope:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, const: true }
        data: {}
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [success, error, request_id]
      properties:
        success: { type: boolean, const: false }
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [AUTH_REQUIRED, AUTH_INVALID, INVALID_CREDENTIALS, SESSION_REFRESH_FAILED, FORBIDDEN, TENANT_REQUIRED, SCHOOL_NOT_FOUND, ACADEMIC_YEAR_NOT_FOUND, CLASS_NOT_FOUND, STUDENT_NOT_FOUND, CARD_NOT_FOUND, DEVICE_NOT_REGISTERED, DEVICE_AUTH_INVALID, DEVICE_NOT_GATE, ATTENDANCE_DISABLED, CARD_BLOCKED, STUDENT_INACTIVE, GATE_RULE_VIOLATION, PARENT_LINK_INVALID, CHILD_NOT_LINKED, RESOURCE_NOT_FOUND, INVALID_JSON, PAYLOAD_TOO_LARGE, VALIDATION_ERROR, CONFLICT, RATE_LIMITED, ROUTE_NOT_FOUND, DATABASE_ERROR, SERVER_UNAVAILABLE, INTERNAL_ERROR]
            message: { type: string }
            details: {}
        request_id: { type: string }
    PaginationMeta:
      type: object
      required: [page, page_size, total]
      properties:
        page: { type: integer }
        page_size: { type: integer }
        total: { type: integer }
...
    DeviceType: { type: string, enum: [GATE, LIBRARY, LED, CARD_STATION, WASTE_SCALE, OTHER] }
...
    DeviceInventoryItem:
      type: object
      required: [id, school_id, device_code, device_type, name, status]
      properties:
        id: { type: string, format: uuid }
        school_id: { type: string, format: uuid }
        device_code: { type: string }
        device_type: { $ref: '#/components/schemas/DeviceType' }
        name: { type: string }
        location: { type: [string, 'null'] }
        firmware_version: { type: [string, 'null'] }
        hardware_version: { type: [string, 'null'] }
        update_channel: { type: string }
        status: { type: string, enum: [PROVISIONING, ACTIVE, DISABLED, MAINTENANCE, RETIRED] }
        last_seen_at: { type: [string, 'null'], format: date-time }
    DeviceListResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data: { type: array, items: { $ref: '#/components/schemas/DeviceInventoryItem' } }
            meta: { $ref: '#/components/schemas/PaginationMeta' }
...